+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [1.9.1](https://github.com/samtun/hackworld/compare/v1.9.0...v1.9.1) (2026-01-03)
+
+
+### Bug Fixes
+
+* add safer validation in updateParticleScaleFactor ([2a8b80c](https://github.com/samtun/hackworld/commit/2a8b80c01f4b49c487ba136c427df35bd393168b))
+* make particle sizes screen-size independent ([7c9a56b](https://github.com/samtun/hackworld/commit/7c9a56b0474091b6a2af14c34bc7b064c7a51bca))
+
 # [1.9.0](https://github.com/samtun/hackworld/compare/v1.8.0...v1.9.0) (2026-01-03)